move data to internal folder
